requirements.txt
beautifulsoup4
python-dotenv
google-generativeai
lxml
requests
playwright